% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFUNCTIONS.R
\name{calcDRIC}
\alias{calcDRIC}
\title{Function to calculate the value of Double Robust Information Criterion (DRIC)}
\usage{
calcDRIC(out.penalized.G, n, ni, p, y, a, E.a, l, wc.str)
}
\arguments{
\item{out.penalized.G}{A list corresponding to the output of penalized G-estimation for a 
single value of the tuning parameter. This is the output of the function penG().}

\item{n}{A scalar value indicating the number of subjects.}

\item{ni}{A list of length n containing the number of measurements per subject.}

\item{p}{A scalar value showing the total number of main effects to estimate (including the intercept).}

\item{y}{A list of length n, where each item in the list is a vector of dimension ni*1 containing the responses from a subject.}

\item{a}{A list of length n, where each item in the list is a vector of dimension ni*1 containing the exposure status of a subject.}

\item{E.a}{A list of length n, where each item in the list is a vector of dimension ni*1 containing the propensity scores of a subject.}

\item{l}{A list of length n, where each item in the list is a matrix of dimension ni*p containing the covariate values for a subject.}

\item{wc.str}{A working correlation structure.}
}
\value{
The value of DRIC under a working correlation structure
}
\description{
Function to calculate the value of Double Robust Information Criterion (DRIC)
}
