% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFUNCTIONS.R
\name{corMatF}
\alias{corMatF}
\title{Function to construct the correlation matrix for a specific subject}
\usage{
corMatF(alpha, ni, corstr)
}
\arguments{
\item{alpha}{The value of correlation parameter (a scalar value for "exchangeable" or "AR(1)" 
structure, or a vector with the elements of the upper triangular part for "unstructured").}

\item{ni}{A scalar value showing the number of measurements for a subject.}

\item{corstr}{The working correlation structure.}
}
\value{
The correlation matrix of dimension ni*ni.
}
\description{
Function to construct the correlation matrix for a specific subject
}
