% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFUNCTIONS.R
\name{corr.ar1.fun}
\alias{corr.ar1.fun}
\title{Function to perform method of moments estimation for the correlation parameter alpha under AR(1) structure}
\usage{
corr.ar1.fun(n, ni, e, phi)
}
\arguments{
\item{n}{A scalar value indicating the number of subjects.}

\item{ni}{A list of length n containing the number of measurements per subject.}

\item{e}{A list of length n containing the residuals.}

\item{phi}{The estimate of variance parameter.}
}
\value{
Method of moments estimate for the correlation parameter alpha under AR(1) structure
}
\description{
Function to perform method of moments estimation for the correlation parameter alpha under AR(1) structure
}
