% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFUNCTIONS.R
\name{phi.hat.fun}
\alias{phi.hat.fun}
\title{Function to perform method of moments estimation for the variance parameter}
\usage{
phi.hat.fun(n, ni, e)
}
\arguments{
\item{n}{A scalar value indicating the number of subjects.}

\item{ni}{A list of length n containing the number of measurements per subject.}

\item{e}{A list of length n containing the residuals.}
}
\value{
Method of moments estimate for the variance parameter
}
\description{
Function to perform method of moments estimation for the variance parameter
}
