% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFUNCTIONS.R
\name{q.mcp}
\alias{q.mcp}
\title{Function to construct the derivative function of MCP penalty}
\usage{
q.mcp(lambda, par)
}
\arguments{
\item{lambda}{A scalar value of the tuning parameter.}

\item{par}{A scalar value for a single regression parameter estimate}
}
\value{
The derivative function of MCP penalty
}
\description{
Function to construct the derivative function of MCP penalty
}
