% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFUNCTIONS.R
\name{q.scad}
\alias{q.scad}
\title{Function to construct the derivative function of SCAD penalty}
\usage{
q.scad(lambda, par)
}
\arguments{
\item{lambda}{A scalar value of the tuning parameter.}

\item{par}{A scalar value for a single regression parameter estimate}
}
\value{
The derivative function of SCAD penalty
}
\description{
Function to construct the derivative function of SCAD penalty
}
