% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFUNCTIONS.R
\name{vec2uMat}
\alias{vec2uMat}
\title{Function to construct correlation matrix from correlation parameters for unstructured correlation structure}
\usage{
vec2uMat(alpha, ni)
}
\arguments{
\item{alpha}{A vector with the elements of the upper triangular part for "unstructured".}

\item{ni}{A scalar value showing the number of measurements for a subject.}
}
\value{
An unstructured correlation matrix
}
\description{
Function to construct correlation matrix from correlation parameters for unstructured correlation structure
}
