% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFUNCTIONS.R
\name{penG}
\alias{penG}
\title{Function to perform penalized G-estimation for a single value of tuning parameter}
\usage{
penG(data, wc.str, tf.model, treat.model, lambda, maxitr, penalty)
}
\arguments{
\item{data}{A data frame containing the variables in longitudinal format. The name of the
variables ID, exposure and outcome must be "id", "a" and "y", respectively, in the data.}

\item{wc.str}{A character string specifying the working correlation structure. The 
following are currently allowed: "independence", "exchangeable", "ar1", and "unstructured".}

\item{tf.model}{A single formula object specifying the covariates of a (linear)
treatment-free model.}

\item{treat.model}{A single formula object specifying the covariates of the logistic
regression model for the binary treatment.}

\item{lambda}{A single value of the tuning parameter.}

\item{maxitr}{Maximum number of iterations allowed.}

\item{penalty}{The penalty type to be used, available options include "SCAD" and "MCP". 
"SCAD" refers to the Smoothly Clipped Absolute Deviation penalty and "MCP" refers to the Minimax Concave Penalty.}
}
\value{
A list containing the estimates, asymptotic variance etc.
}
\description{
This function performs penalized G-estimation for a given longitudinal data,
a specific working correlation structure and a single value of the tuning parameter.
}
